+++ src/QuickMarcEditor/QuickMarcEditorRows/QuickMarcEditorRows.js
@@ -34,16 +34,18 @@ import { DeletedRowPlaceholder } from './DeletedRowPlaceholder';
 import { LinkButton } from './LinkButton';
 import { SplitField } from './SplitField';
 import {
-  isReadOnly,
   hasIndicatorException,
   hasAddException,
-  hasDeleteException,
   hasMoveException,
+} from './utils';
+import {
+  checkCanBeLinked,
+  isFixedFieldRow,
   isMaterialCharsRecord,
   isPhysDescriptionRecord,
-  isFixedFieldRow,
-} from './utils';
-import { checkCanBeLinked } from '../utils';
+  isReadOnly,
+  hasDeleteException,
+} from '../utils';
 import { useAuthorityLinking } from '../../hooks';
 import { QUICK_MARC_ACTIONS } from '../constants';
 import {
@@ -68,6 +70,7 @@ const QuickMarcEditorRows = ({
     restoreRecord,
   },
   marcType,
+  instance,
 }) => {
   const stripes = useStripes();
   const intl = useIntl();
@@ -210,7 +213,7 @@ const QuickMarcEditorRows = ({
             const isDisabled = isReadOnly(recordRow, action, marcType);
             const withIndicators = !hasIndicatorException(recordRow);
             const withAddRowAction = hasAddException(recordRow, marcType);
-            const withDeleteRowAction = hasDeleteException(recordRow, marcType);
+            const withDeleteRowAction = hasDeleteException(recordRow, marcType, instance, initialValues);
             const withMoveUpRowAction = hasMoveException(recordRow, fields[idx - 1]);
             const withMoveDownRowAction = hasMoveException(recordRow, fields[idx + 1]);
 
@@ -479,6 +482,7 @@ const QuickMarcEditorRows = ({
 
 QuickMarcEditorRows.propTypes = {
   action: PropTypes.oneOf(Object.values(QUICK_MARC_ACTIONS)).isRequired,
+  instance: PropTypes.object,
   type: PropTypes.string.isRequired,
   subtype: PropTypes.string.isRequired,
   fields: PropTypes.arrayOf(PropTypes.shape({