+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 * [UIQM-373](https://issues.folio.org/browse/UIQM-373) Save buttons are not disabled after removing unsaved updates from linked field in "MARC bib" record
 * [UIQM-376](https://issues.folio.org/browse/UIQM-376) Reference MARC Authority record is opened when user clicks on the "MARC Authority" icon next to the controlled field
 * [UIQM-349](https://issues.folio.org/browse/UIQM-349) Align the module with mod-search API breaking change
+* [UIQM-348](https://issues.folio.org/browse/UIQM-348) Change to Linked MARC authority record has been updated confirmation messaging
 
 ## [5.2.0](https://github.com/folio-org/ui-quick-marc/tree/v5.2.0) (2022-10-26)